refactor: exchange models in exchange app and only migrate exchange
