+++ django/university/routes/exchange/options/ExchangeOptionsView.py
@@ -5,8 +5,9 @@ from django.http import HttpResponse, JsonResponse
 from django.db.models import Prefetch
 
 from university.exchange.utils import curr_semester_weeks
-from university.models import MarketplaceExchange, MarketplaceExchangeClass
 from university.serializers.MarketplaceExchangeClassSerializer import MarketplaceExchangeClassSerializer
+
+from exchange.models import MarketplaceExchange, MarketplaceExchangeClass
 
 class ExchangeOptionsView(APIView):
     """
