+++ docker-compose.yaml
@@ -1,36 +1,34 @@
-version: "3.7"
-
 services:
-  db:
-    build: ./postgres
-    container_name: tts_db
-    volumes:
-      - postgres_data:/var/lib/postgresql/data/
-    env_file:
-      - postgres/.env.dev
-    ports:
-      - 5432:5432
-    ulimits:
-      nofile:
-        soft: 262144
-        hard: 262144
+  # db:
+  #   build: ./postgres
+  #   container_name: tts_db
+  #   volumes:
+  #     - postgres_data:/var/lib/postgresql/data/
+  #   env_file:
+  #     - postgres/.env.dev
+  #   ports:
+  #     - 5432:5432
+  #   ulimits:
+  #     nofile:
+  #       soft: 262144
+  #       hard: 262144
 
-  pgadmin:
-    build:
-      context: ./pgadmin
-    container_name: tts_pgadmin
-    env_file:
-      - pgadmin/.env.dev
-    volumes:
-      - pgadmin_data:/var/lib/pgadmin
-    ports:
-      - 4000:80
+  # pgadmin:
+  #   build:
+  #     context: ./pgadmin
+  #   container_name: tts_pgadmin
+  #   env_file:
+  #     - pgadmin/.env.dev
+  #   volumes:
+  #     - pgadmin_data:/var/lib/pgadmin
+  #   ports:
+  #     - 4000:80
   
   nginx:
     image: nginx:1-alpine
     container_name: tts_nginx
     ports:
-      - 443:443
+      - 4431:443
     volumes:
       - ./nginx/nginx.conf:/etc/nginx/nginx.conf
       - ./nginx/certs/:/etc/nginx/certs/
